requirements.txt
gradio
Pillow
openai==0.28
python-dotenv
pandas